requirements.txt
django
djangorestframework
flask
requests
streamlit
psycopg2-binary